Support for explicit ordering of POJO fields
A protobuf message requires the fields to be numbered. Up to now
we have relied on the (unreliable and undefined) ordering of
declared fields. It's probably OK to keep that as a default because
it's super convenient, but we want users to be able to explicitly
order the fields in a POJO. You can no do that with @Order
or @Priority. You can even omit the annotation and fall back to
the declaration order as long as all fields get a unique order.